refactor: changing the tests directory and putting the non authenticated route in a different line
