+++ src/main/java/br/ufcg/animais/animais_ufcg/infra/security/SecurityConfig.java
@@ -31,7 +31,8 @@ public class SecurityConfig {
                 .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
                 .authorizeHttpRequests(authorize -> authorize
                         .requestMatchers(HttpMethod.POST, "/auth/login", "auth/register").permitAll()
-                        .requestMatchers(HttpMethod.GET, "/h2-console", "/animal/getAvailable").permitAll()
+                        .requestMatchers(HttpMethod.GET, "/h2-console").permitAll()
+                        .requestMatchers(HttpMethod.GET, "/animal/getAvailable").permitAll()
                         .anyRequest().authenticated()
                 )
                 .addFilterBefore(securityFilter, UsernamePasswordAuthenticationFilter.class);
